requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.1
djangorestframework==3.14.0
idna==3.6
iniconfig==2.0.0
packaging==23.2
pluggy==1.4.0
pytest==7.4.4
pytest-django==4.7.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.1.0

Faker~=22.5.1